write a really early version of the global config
This has a few purposes. First, it defines some convenience functions
and macros that most users will want to have around. Second, it
demonstrates the general semantics of the DSL we'll use for binding
actions to joystick functionality. Third, I'm using it as a really
crude version of TDD. :P