switch to new binding organization
hold procedures bound to buttons inside C code, attach new ones using
bind-button. this is a change from before, where you'd build up a giant
alist of mappings and then feed it to the event loop. this feels a lot
cleaner and nicer from the user's perspective. also more efficient,
since now we store procedures indexed by their button and action and
have a linked list that we just traverse. this will only matter if you
have thousands of procedures bound, but it's nice a clean. :)