big swath of new features.
joystick.c handles axes value normalization and history now. this should
be a relatively large performance boost, and that seems to be the case;
jitter and lag are way down again. built a bunch of new stuff in the
scheme half of the extension. axes bindings are better now. tore the
installation stuff out of make because I realized it was
broken. provided a "make test" makefile target that automatically sets
up and runs code.